@@ -1,22 +1,23 @@
 name: standard_dialogs
-version: 1.2.2
+version: 2.0.0
 description: Package with standardized dialogues for use in your application or package.
 homepage: https://github.com/ricardocrescenti
 repository: https://github.com/ricardocrescenti/flutter_standard_dialogs
 issue_tracker: https://github.com/ricardocrescenti/flutter_standard_dialogs/issues
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  simple_localization: ^1.2.2
+
+  simple_localization: ^2.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.8.0
+  flutter_lints: ^2.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
