@@ -7,54 +7,54 @@ class ResultDialog extends StatelessWidget {
 	final Color backgroundColor;
 	final Color textColor;
 	final Widget icon;
-	final Widget title;
-	final Widget content;
+	final Widget? title;
+	final Widget? content;
 	final DialogAction action;
 	final Radius borderRadius = const Radius.circular(5);
 
 	ResultDialog({
-		@required this.backgroundColor,
-		@required this.textColor,
-		@required this.icon,
+		Key? key,
+		required this.backgroundColor,
+		required this.textColor,
+		required this.icon,
 		this.title,
 		this.content,
-		this.action
-	}) {
+		required this.action
+	}) : super(key: key) {
 		assert(title != null || content != null);
 	}
 
 	@override
 	Widget build(BuildContext context) {
 		
-		return Container(
-			child: AlertDialog(
-				title: buildTitle(context),
-				titlePadding: EdgeInsets.zero,
-				contentPadding: EdgeInsets.zero,
-				content: Container(
-					decoration: BoxDecoration(
-						color: Theme.of(context).scaffoldBackgroundColor,
-						borderRadius: BorderRadius.only(
-								bottomLeft: borderRadius,
-								bottomRight: borderRadius
-						)
-					),
-					child: Column(
-						crossAxisAlignment: CrossAxisAlignment.stretch,
-						mainAxisSize: MainAxisSize.min,
-						children: [
-							buildContend(context),
-							buildActions(context)
-						]
+		return AlertDialog(
+			title: buildTitle(context),
+			titlePadding: EdgeInsets.zero,
+			contentPadding: EdgeInsets.zero,
+			content: Container(
+				decoration: BoxDecoration(
+					color: Theme.of(context).scaffoldBackgroundColor,
+					borderRadius: BorderRadius.only(
+							bottomLeft: borderRadius,
+							bottomRight: borderRadius
 					)
 				),
-				backgroundColor: Colors.transparent,
-			)
+				child: Column(
+					crossAxisAlignment: CrossAxisAlignment.stretch,
+					mainAxisSize: MainAxisSize.min,
+					children: [
+						buildContend(context),
+						buildActions(context)
+					]
+				)
+			),
+			backgroundColor: Colors.transparent,
 		);
 
 	}
 
 	buildTitle(BuildContext context) {
+
 		return Container(
 			decoration: BoxDecoration(
 				color: backgroundColor,
@@ -64,11 +64,11 @@ class ResultDialog extends StatelessWidget {
 				)
 			),
 			//color: color,
-			padding: EdgeInsets.symmetric(vertical: 40, horizontal: 15),
+			padding: const EdgeInsets.symmetric(vertical: 40, horizontal: 15),
 			child: Column(
 				children: [
 					Padding(
-						padding: EdgeInsets.only(bottom: 10),	
+						padding: const EdgeInsets.only(bottom: 10),	
 						child: (icon is Icon 
 							? IconTheme.merge(
 								data: IconThemeData(size: 70, color: textColor), 
@@ -79,20 +79,21 @@ class ResultDialog extends StatelessWidget {
 					(title != null 
 						? (title is Text 
 							? AnimatedDefaultTextStyle(
-								style: (title as Text).style ?? Theme.of(context).textTheme.headline6.copyWith(color: textColor),
+								style: (title as Text).style ?? Theme.of(context).textTheme.headline6!.copyWith(color: textColor),
 								textAlign: (title as Text).textAlign ?? TextAlign.center,
 								duration: kThemeChangeDuration,
-								child: title)
+								child: title!)
 							: title) 
 						: Container()
-					)
+					)!
 				],
 			),
 		);
+
 	}
 
 	buildContend(BuildContext context) {
-		if (this.content == null) {
+		if (content == null) {
 			return Container();
 		}
 
@@ -100,10 +101,10 @@ class ResultDialog extends StatelessWidget {
 			padding: const EdgeInsets.fromLTRB(24, 20, 24, 0),
 			child: (content is Text 
 				? AnimatedDefaultTextStyle(
-					style: (content as Text).style ?? Theme.of(context).textTheme.subtitle1,
+					style: (content as Text).style ?? Theme.of(context).textTheme.subtitle1!,
 					duration: kThemeChangeDuration,
 					textAlign: (content as Text).textAlign ?? TextAlign.center,
-					child: this.content)
+					child: content!)
 				: content
 			),
 		);
@@ -111,21 +112,21 @@ class ResultDialog extends StatelessWidget {
 
 	buildActions(BuildContext context) {
 		return Container(
-			padding: EdgeInsets.symmetric(vertical: 15),
+			padding: const EdgeInsets.symmetric(vertical: 15),
 			child: Row(
 				mainAxisAlignment: MainAxisAlignment.center,
 				children: [
 					ElevatedButton(
-						child: Padding(
-							padding: EdgeInsets.symmetric(horizontal: 25),
-							child: action.title
-						),
 						style: ElevatedButton.styleFrom(
 							primary: backgroundColor,
 							onPrimary: textColor,
 							shape: RoundedRectangleBorder(
 								borderRadius: BorderRadius.circular(30.0)
 							)
+						),
+						child: Padding(
+							padding: const EdgeInsets.symmetric(horizontal: 25),
+							child: action.title
 						),
 						onPressed: () => action.performClick(context)
 					)
